Additional Combobox for "select building" & typo fixed
with additional documentation
