+++ src/main/java/org/jevis/jeconfig/structurewizard/ManualWizardStep2.java
@@ -40,7 +40,7 @@ import org.jevis.api.JEVisAttribute;
 import org.jevis.api.JEVisClass;
 import org.jevis.api.JEVisException;
 import org.jevis.api.JEVisObject;
-import org.jevis.jeconfig.bulgedit.CreateTable;
+import org.jevis.jeconfig.bulkedit.CreateTable;
 import org.jevis.jeconfig.structurewizard.WizardSelectedObject;
 import org.jevis.jeconfig.plugin.object.ObjectTree;
 import org.jevis.jeconfig.tool.ImageConverter;
